requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
groq==0.8.0
typing-extensions==4.8.0
speech-recognition==3.10.0
pydub==0.25.1
